feat: app hosting release (#400)
* feat: cca integrate upload scripts (#392)

Adds upload and upload-ci to the cca template.
Explains the commands in the readme of the template

* chore: update version of app scripts in template (#402)

* feat(Caching): cache Org Id, access token & app def id from create-definition process

* feat(Caching): move all key names to constants file

* chore: Remove duplicate get-mgmt-token

* feat(Caching): Use Cached env vars for uploading app upload

* - Start with cache first before showing prompt
- Minor UI/Text improvements

* Change default folder bundle folder to ./build

* Minor Text alignment

* Minor logging and readability fixes

* fix: set default location for activating bundle when no location is set

* feat: [] Automated canary release (#421)

add create canary release to ci config and as script

* feat: [] adds `open-settings` command to app-scripts (#422)

Adds a simple open command to app-scripts.
The structure of the call is a bit different as there is no sense to have an interactive/non-interactive mode, as this command will only run in one place and never in a CI.

So there is only one function executed.

* feat: [] adds `bundle-cleanup` command to app-scripts (#428)

Adds the bundle-cleanup to the app-scripts.

Added also the optional --keep parameter to adjust the number of bundles to keep instead of hard-coded 50. (50 is still the default when not provided)

* feat: cca integrate upload scripts (#392)

Adds upload and upload-ci to the cca template.
Explains the commands in the readme of the template

* chore: update version of app scripts in template (#402)

* feat(Caching): cache Org Id, access token & app def id from create-definition process

* feat(Caching): move all key names to constants file

* chore: Remove duplicate get-mgmt-token

* feat(Caching): Use Cached env vars for uploading app upload

* - Start with cache first before showing prompt
- Minor UI/Text improvements

* Change default folder bundle folder to ./build

* Minor Text alignment

* Minor logging and readability fixes

* fix: set default location for activating bundle when no location is set

* feat: [] Automated canary release (#421)

add create canary release to ci config and as script

* feat: [] adds `open-settings` command to app-scripts (#422)

Adds a simple open command to app-scripts.
The structure of the call is a bit different as there is no sense to have an interactive/non-interactive mode, as this command will only run in one place and never in a CI.

So there is only one function executed.

* feat: [] adds `bundle-cleanup` command to app-scripts (#428)

Adds the bundle-cleanup to the app-scripts.

Added also the optional --keep parameter to adjust the number of bundles to keep instead of hard-coded 50. (50 is still the default when not provided)

* fix: update lock file

* fix: add dev dependencies to manifest creation (#513)

With app-scripts being part of the devDependencies the outdated check should also include the devDependencies

* feat: run dependency check on canary branch too (#514)

Also added the manually updated template.json file in the template package

* doc: update link in readme

Co-authored-by: Kado Damball <kadodamball@hotmail.com>
Co-authored-by: kdamball <kdamball@users.noreply.github.com>
Co-authored-by: Manuel Spagnolo <spagnolo.manu@gmail.com>